+++ src/vmm/src/device_manager/persist.rs
@@ -154,6 +154,7 @@ pub struct ConnectedLegacyState {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct MmdsState {
     version: MmdsVersion,
+    imds_compat: bool,
 }
 
 /// Holds the device states.
@@ -326,8 +327,10 @@ impl<'a> Persist<'a> for MMIODeviceManager {
                 TYPE_NET => {
                     let net = locked_device.as_any().downcast_ref::<Net>().unwrap();
                     if let (Some(mmds_ns), None) = (net.mmds_ns.as_ref(), states.mmds.as_ref()) {
+                        let mmds_guard = mmds_ns.mmds.lock().expect("Poisoned lock");
                         states.mmds = Some(MmdsState {
-                            version: mmds_ns.mmds.lock().expect("Poisoned lock").version(),
+                            version: mmds_guard.version(),
+                            imds_compat: mmds_guard.imds_compat(),
                         });
                     }
 
@@ -539,7 +542,11 @@ impl<'a> Persist<'a> for MMIODeviceManager {
         if let Some(mmds) = &state.mmds {
             constructor_args
                 .vm_resources
-                .set_mmds_basic_config(mmds.version, constructor_args.instance_id)?;
+                .set_mmds_basic_config(
+                    mmds.version,
+                    mmds.imds_compat,
+                    constructor_args.instance_id,
+                )?;
         }
 
         for net_state in &state.net_devices {
@@ -750,6 +757,7 @@ mod tests {
                 &mut event_manager,
                 network_interface,
                 MmdsVersion::V2,
+                true,
             );
             // Add a vsock device.
             let vsock_dev_id = "vsock";
@@ -826,7 +834,8 @@ mod tests {
     "network_interfaces": [
       "netif"
     ],
-    "ipv4_address": "169.254.169.254"
+    "ipv4_address": "169.254.169.254",
+    "imds_compat": true
   }},
   "network-interfaces": [
     {{
@@ -859,7 +868,18 @@ mod tests {
                 .version(),
             MmdsVersion::V2
         );
-        assert_eq!(device_states.mmds.unwrap().version, MmdsVersion::V2.into());
+        assert!(
+            vm_resources
+                .mmds
+                .as_ref()
+                .unwrap()
+                .lock()
+                .unwrap()
+                .imds_compat()
+        );
+        let mmds_state = device_states.mmds.unwrap();
+        assert_eq!(mmds_state.version, MmdsVersion::V2.into());
+        assert!(mmds_state.imds_compat);
 
         assert_eq!(restored_dev_manager, original_mmio_device_manager);
         assert_eq!(